+++ tests/Highcharts/ExportingTest.php
@@ -18,32 +18,33 @@ use HighchartsBundle\Tests\AbstractChartTestCase;
 
 /**
  * This class hold Unit Tests for the exporting option.
+ *
+ * @extends AbstractChartTestCase<Highchart>
  */
 final class ExportingTest extends AbstractChartTestCase
 {
     /**
-     * buttons option.
+     * Buttons option.
      */
     public function testButtons(): void
     {
-        $chart = new Highchart();
         // align option (string - left/center/right)
-        $chart->exporting['buttons'] = ['exportButton' => ['align' => 'center']];
+        $this->chart->exporting['buttons'] = ['exportButton' => ['align' => 'center']];
         $regex = '/exporting: \{"buttons":\{"exportButton":\{"align":"center"\}\}\}/';
-        self::assertChartMatchesRegularExpression($chart, $regex);
+        self::assertChartMatchesRegularExpression($this->chart, $regex);
 
-        $chart->exporting['buttons'] = ['printButton' => ['align' => 'center']];
+        $this->chart->exporting['buttons'] = ['printButton' => ['align' => 'center']];
         $regex = '/exporting: \{"buttons":\{"printButton":\{"align":"center"\}\}\}/';
-        self::assertChartMatchesRegularExpression($chart, $regex);
+        self::assertChartMatchesRegularExpression($this->chart, $regex);
 
         // backgroundColor option
-        $chart->exporting['buttons'] = ['exportButton' => ['backgroundColor' => 'blue']];
+        $this->chart->exporting['buttons'] = ['exportButton' => ['backgroundColor' => 'blue']];
         $regex = '/exporting: \{"buttons":\{"exportButton":\{"backgroundColor":"blue"\}\}\}/';
-        self::assertChartMatchesRegularExpression($chart, $regex);
+        self::assertChartMatchesRegularExpression($this->chart, $regex);
 
-        $chart->exporting['buttons'] = ['printButton' => ['backgroundColor' => 'blue']];
+        $this->chart->exporting['buttons'] = ['printButton' => ['backgroundColor' => 'blue']];
         $regex = '/exporting: \{"buttons":\{"printButton":\{"backgroundColor":"blue"\}\}\}/';
-        self::assertChartMatchesRegularExpression($chart, $regex);
+        self::assertChartMatchesRegularExpression($this->chart, $regex);
     }
 
     /**
@@ -51,69 +52,70 @@ final class ExportingTest extends AbstractChartTestCase
      */
     public function testEnabled(): void
     {
-        $chart = new Highchart();
-        $chart->exporting['enabled'] = true;
+        $this->chart->exporting['enabled'] = true;
         $regex = '/exporting: \{"enabled":true\}/';
-        self::assertChartMatchesRegularExpression($chart, $regex);
+        self::assertChartMatchesRegularExpression($this->chart, $regex);
 
-        $chart->exporting['enabled'] = false;
+        $this->chart->exporting['enabled'] = false;
         $regex = '/exporting: \{"enabled":false\}/';
-        self::assertChartMatchesRegularExpression($chart, $regex);
+        self::assertChartMatchesRegularExpression($this->chart, $regex);
     }
 
     /**
-     * filename option (string).
+     * Filename option (string).
      */
     public function testFilename(): void
     {
-        $chart = new Highchart();
-        $chart->exporting['filename'] = 'graph';
+        $this->chart->exporting['filename'] = 'graph';
         $regex = '/exporting: \{"filename":"graph"\}/';
-        self::assertChartMatchesRegularExpression($chart, $regex);
+        self::assertChartMatchesRegularExpression($this->chart, $regex);
     }
 
     /**
-     * type option (string - image/png, image/jpeg, application/pdf or image/svg+xml).
+     * Type option (string - image/png, image/jpeg, application/pdf or image/svg+xml).
      */
     public function testType(): void
     {
-        $chart = new Highchart();
-        $chart->exporting['type'] = 'image/png';
+        $this->chart->exporting['type'] = 'image/png';
         $regex = '/exporting: \{"type":"image\/png"\}/';
-        self::assertChartMatchesRegularExpression($chart, $regex);
+        self::assertChartMatchesRegularExpression($this->chart, $regex);
 
-        $chart->exporting['type'] = 'image/jpeg';
+        $this->chart->exporting['type'] = 'image/jpeg';
         $regex = '/exporting: \{"type":"image\/jpeg"\}/';
-        self::assertChartMatchesRegularExpression($chart, $regex);
+        self::assertChartMatchesRegularExpression($this->chart, $regex);
 
-        $chart->exporting['type'] = 'application/pdf';
+        $this->chart->exporting['type'] = 'application/pdf';
         $regex = '/exporting: \{"type":"application\/pdf"\}/';
-        self::assertChartMatchesRegularExpression($chart, $regex);
+        self::assertChartMatchesRegularExpression($this->chart, $regex);
 
-        $chart->exporting['type'] = 'image/svg+xml';
+        $this->chart->exporting['type'] = 'image/svg+xml';
         $regex = '/exporting: \{"type":"image\/svg\+xml"\}/';
-        self::assertChartMatchesRegularExpression($chart, $regex);
+        self::assertChartMatchesRegularExpression($this->chart, $regex);
     }
 
     /**
-     * url option (string).
+     * Url option (string).
      */
     public function testUrl(): void
     {
-        $chart = new Highchart();
-        $chart->exporting['url'] = 'https://www.google.com';
+        $this->chart->exporting['url'] = 'https://www.google.com';
         $regex = '/exporting: \{"url":"https:\/\/www.google.com"\}/';
-        self::assertChartMatchesRegularExpression($chart, $regex);
+        self::assertChartMatchesRegularExpression($this->chart, $regex);
     }
 
     /**
-     * width option (integer - width in px).
+     * Width option (integer - width in px).
      */
     public function testWidth(): void
     {
-        $chart = new Highchart();
-        $chart->exporting['width'] = 300;
+        $this->chart->exporting['width'] = 300;
         $regex = '/exporting: \{"width":300\}/';
-        self::assertChartMatchesRegularExpression($chart, $regex);
+        self::assertChartMatchesRegularExpression($this->chart, $regex);
+    }
+
+    #[\Override]
+    protected function createChart(): Highchart
+    {
+        return new Highchart();
     }
 }