+++ tests/Highcharts/SeriesTest.php
@@ -18,6 +18,8 @@ use HighchartsBundle\Tests\AbstractChartTestCase;
 
 /**
  * This class hold Unit Tests for the series option.
+ *
+ * @extends AbstractChartTestCase<Highchart>
  */
 final class SeriesTest extends AbstractChartTestCase
 {
@@ -31,13 +33,18 @@ final class SeriesTest extends AbstractChartTestCase
             ['name' => 'Data #2', 'data' => [7, 3, 5, 1, 6, 5, 9]],
         ];
 
-        $chart = new Highchart();
-        $chart->series->merge($series);
+        $this->chart->series->merge($series);
 
         $regex = '/\{"name":"Data #1","data":\[1,2,4,5,6,3,8\]\}/';
-        self::assertChartMatchesRegularExpression($chart, $regex);
+        self::assertChartMatchesRegularExpression($this->chart, $regex);
 
         $regex = '/\{"name":"Data #2","data":\[7,3,5,1,6,5,9\]\}/';
-        self::assertChartMatchesRegularExpression($chart, $regex);
+        self::assertChartMatchesRegularExpression($this->chart, $regex);
+    }
+
+    #[\Override]
+    protected function createChart(): Highchart
+    {
+        return new Highchart();
     }
 }