+++ tests/Highstock/ColorsTest.php
@@ -16,14 +16,22 @@ namespace HighchartsBundle\Tests\Highstock;
 use HighchartsBundle\Highcharts\Highstock;
 use HighchartsBundle\Tests\AbstractChartTestCase;
 
+/**
+ * @extends AbstractChartTestCase<Highstock>
+ */
 final class ColorsTest extends AbstractChartTestCase
 {
     public function testColors(): void
     {
-        $chart = new Highstock();
         $colors = ['#FF0000', '#00FF00', '#0000FF'];
-        $chart->colors = $colors;
+        $this->chart->colors = $colors;
         $regex = '/colors: \["#FF0000","#00FF00","#0000FF"\]/';
-        self::assertChartMatchesRegularExpression($chart, $regex);
+        self::assertChartMatchesRegularExpression($this->chart, $regex);
+    }
+
+    #[\Override]
+    protected function createChart(): Highstock
+    {
+        return new Highstock();
     }
 }